requirements.txt
blinker==1.9.0
boto3==1.42.50
botocore==1.42.50
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
Flask==3.1.2
flask-cors==6.0.1
google-auth==2.41.1
google-auth-oauthlib==1.2.3
gunicorn==23.0.0
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.1.0
lxml==6.0.2
MarkupSafe==3.0.3
oauthlib==3.3.1
packaging==26.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pymupdf==1.26.5
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
requests==2.32.5
requests-oauthlib==2.0.0
reportlab>=4.0.0
rsa==4.9.1
scikit-learn==1.6.1
s3transfer==0.16.0
six==1.17.0
typing_extensions==4.15.0
urllib3==1.26.20
Werkzeug==3.1.3
zipp==3.23.0
